backend/requirements.txt
fastapi>=0.115.11
uvicorn>=0.34.0
redis>=5.2.1
python-dotenv>=1.0.1
pydantic>=2.10.6
pydantic-settings>=2.8.1
pytest>=8.3.5
httpx>=0.28.1
tiktoken
gitingest
requests==2.32.3
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-generativeai==0.3.2
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
proto-plus==1.26.1
protobuf==4.25.6
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-multipart==0.0.6
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
pinecone-client==3.0.0
retry